Compare human readable key codes with chars literals.
Many sequences are actually human readable and fall into the printable
ASCII subset, they are a lot more recognizable when written as chars
compared to numbers.

For example up arrow is sent back from the terimal as ESC [A, and so
forth.

This commit makes the code speak the same "language" that you find in
any terminal escape sequences documentation.