Clarifying Free Software vs 'Commercial'
Free Software is not incompatible at all with commercial use. Indeed, quite the opposite; were I running a company based on open code, it would be significantly to my advantage to choose a copyleft license. But that's beside the point.

Free Software, then, is not disjunct from Commercial, and it would be more appropriate to say that the BSD permits use in Free Software or *Proprietary* use-cases, because that's the real difference between, say, the GPL and BSD.