Enable tab completion in the middle, currently only the end of the line could be completed. This functionality is done by changing:
1. Pass an additional "pos" argument to the linenoiseCompletionCallback
typedef, this will enable user to get cursor position from within the
completion code (in callback function).

2. Change the linenoiseAddCompletion() API and add "posvec" member to
its corresponding linenoiseCompletions structure and completeLine()
function so that we can save the cursor position (after the inserted
word) for each completion option rather than place cursor at the end
of line if a completion in the middle is chosen.

3. Use a slightly complicated completion() code in example.c, to show
how to achieve an much more useful completion callback that can do tab
completion in the middle.