Allow passing a NULL prompt to linenoise()
When a NULL prompt is passed the linenoise() function will assume that a prompt has already been written.  It computes the length of the prompt from cursor position at the time it is called and avoids writing over the prompt when editing.  This allows using linenoise to add line editing capability to a program (e.g. tclsh) which provides a REPL, so prints its own prompts, but does not provide any command line editing.